moved searchbars state from module to global store since its reused by other components
