--- src/store/SearchBar/getters.js
@@ -1,7 +1,0 @@
-const getters = {
-  filteredByKeyword: (state, getters, rootState) => {
-    return rootState.filteredProducts;
-  }
-};
-
-export default getters;
