--- src/store/SearchBar/mutations.js
@@ -1,3 +1,0 @@
-const mutations = {};
-
-export default mutations;
