+++ src/store/index.js
@@ -1,6 +1,5 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import searchbar from './SearchBar/index.js';
 import actions from './actions.js';
 import mutations from './mutations.js';
 import getters from './getters.js';
@@ -91,7 +90,6 @@ export default new Vuex.Store({
   mutations: mutations,
   getters: getters,
   modules: {
-    searchbar: searchbar,
     hamburgerIcon: hamburgerIcon
   }
 });
